# opt:(decoder) use `runtime.slicebytetostring()` for string copy