- cmd
`./bench.py -b "\"^(BenchmarkDecoder_Generic_Sonic|BenchmarkDecoder_Binding_Sonic)$\"" -c`
- result
```
name                      old time/op    new time/op    delta
Decoder_Generic_Sonic-16    69.2µs ± 2%    70.0µs ± 3%   ~     (p=0.190 n=10+10)
Decoder_Binding_Sonic-16    33.2µs ± 2%    33.0µs ± 1%   ~     (p=0.182 n=10+9)

name                      old speed      new speed      delta
Decoder_Generic_Sonic-16   188MB/s ± 2%   186MB/s ± 4%   ~     (p=0.190 n=10+10)
Decoder_Binding_Sonic-16   393MB/s ± 2%   394MB/s ± 2%   ~     (p=0.393 n=10+10)

name                      old alloc/op   new alloc/op   delta
Decoder_Generic_Sonic-16    57.8kB ± 1%    57.9kB ± 1%   ~     (p=0.796 n=10+10)
Decoder_Binding_Sonic-16    28.2kB ± 1%    28.3kB ± 1%   ~     (p=0.670 n=10+10)

name                      old allocs/op  new allocs/op  delta
Decoder_Generic_Sonic-16       723 ± 0%       723 ± 0%   ~     (all equal)
Decoder_Binding_Sonic-16       137 ± 0%       137 ± 0%   ~     (all equal)
```
```
name                      old time/op    new time/op    delta
Decoder_Binding_Sonic-16    32.8µs ± 1%    35.0µs ± 4%  +6.78%  (p=0.000 n=10+10)
Decoder_Generic_Sonic-16    69.5µs ± 2%    70.8µs ± 4%    ~     (p=0.190 n=10+10)

name                      old speed      new speed      delta
Decoder_Binding_Sonic-16   398MB/s ± 1%   373MB/s ± 4%  -6.32%  (p=0.000 n=10+10)
Decoder_Generic_Sonic-16   188MB/s ± 2%   184MB/s ± 3%    ~     (p=0.190 n=10+10)

name                      old alloc/op   new alloc/op   delta
Decoder_Generic_Sonic-16    57.7kB ± 1%    57.9kB ± 1%    ~     (p=0.190 n=10+10)
Decoder_Binding_Sonic-16    28.3kB ± 0%    28.3kB ± 1%    ~     (p=1.000 n=8+10)

name                      old allocs/op  new allocs/op  delta
Decoder_Generic_Sonic-16       723 ± 0%       723 ± 0%    ~     (all equal)
Decoder_Binding_Sonic-16       137 ± 0%       137 ± 0%    ~     (all equal)
```
```
name                      old time/op    new time/op    delta
Decoder_Generic_Sonic-16    69.8µs ± 1%    70.4µs ± 6%   ~     (p=0.912 n=10+10)
Decoder_Binding_Sonic-16    33.2µs ± 5%    33.7µs ± 6%   ~     (p=0.218 n=10+10)

name                      old speed      new speed      delta
Decoder_Generic_Sonic-16   187MB/s ± 1%   185MB/s ± 5%   ~     (p=0.912 n=10+10)
Decoder_Binding_Sonic-16   393MB/s ± 5%   387MB/s ± 5%   ~     (p=0.218 n=10+10)

name                      old alloc/op   new alloc/op   delta
Decoder_Generic_Sonic-16    57.8kB ± 1%    57.9kB ± 1%   ~     (p=0.529 n=10+10)
Decoder_Binding_Sonic-16    28.3kB ± 1%    28.2kB ± 1%   ~     (p=0.123 n=10+10)

name                      old allocs/op  new allocs/op  delta
Decoder_Generic_Sonic-16       723 ± 0%       723 ± 0%   ~     (all equal)
Decoder_Binding_Sonic-16       137 ± 0%       137 ± 0%   ~     (all equal)
```